Don't use file_class
Although python-fuse provides file_class as a convenient mechanism
to work with file objects, in our case it complicates the setup,
because we are doing another level of proxying (using Container
and Storage objects).

As a result, various classes have unnecessary dependencies, the
control flow is more indirect, python-fuse needs to be worked
around (see WildlandFSResolverMixin), and it becomes much harder
to customize behaviour for a single class.

Instead, we can implement a similar proxy ourselves, and use it
both in the storage objects and the top-level WildlandFS.