Use jsonschema library for manifest validation
More information: https://json-schema.org/

Compared to the voluptuous library, the schemas are more verbose.
They are also validation only, no parsing/conversion.

On the plus side, jsonschema is a universal format, with plenty of
tools available. And because the schemas are stored as data, they
can be used in other tools, including third-party ones.