Use '0xcoffee' fingerprint
This is an unambiguous format since it cannot be mistaken for a key
ID.

As a side effect, we can notice that unquoted identifiers like
0xcoffee will be parsed as integers in YAML. Allowing that sounds
like a bad idea, as it makes implementation for harder for everyone
(handling string/int, handling possible overflow). Fortunately, the
JSON Schema requires key fingerprints to be strings.