Use pip for packages
This has some drawbacks compared to using system packages, e.g. we
have to keep them up to date manually, or the build process
requires some compiling.

However, the upside is that the packages are more up to date than
the distro versions (so far, we had issues with python-fuse and
jsonschema), and it's possible to use some packages that are not
packaged by distros at all (I'm planning to use webdavclient3).

Use pip-tools for requirements. That allows for easy version
pinning, updates, and gives some security by including the hashes
in requirements.txt.