cached storage: ensure right file size
For synthetic files, we still need to provide the right size.
Unfortunately, this is necessary, because the kernel will make use
of that information when handling read(). If the size is too small,
the reads will be truncated, if it's too big, in some cases it will
be padded with zeroes.