Syncing documentation and better handling
Added better handling for storage watchers; now if a storage manifest has a watcher-delay
param, the SimpleStorageWatcher (brute force that scans the storage every watcher-delay seconds)
will be used.

Furthermore, removed a lock from fs.py that could lead to deadlocks. It's safe to remove because
the storage_id is removed from the storage_watches dictionary only ofter the watcher mechanism
is stopped and all events are handled (there's a thread.join in the stop_watcher method).

Also more documentation.