Do not write default parameters into manifests
This skips weird entries like 'title: null' in manifests. But more importantly,
allows to distinguish explicit value set, from default value. And allows
changing that default when necessary.

This in practice is only approximation of what is described above, because it
skips "empty" not "default" values. This is because in click it is impossible
to distinguish not provided parameter (where click fills in default value) from
that the default value provided explicitly.