Allow StorageBackend to be mounted multiple times
Add an internal mount counter to StorageBackend, then add
request_mount/request_unmount calls for the external callers.
This allows the same StorageBackend instance to be "mounted" multiple times and
unmounted, while the actual mount happens only on the first call, and unmount
happens on the last one.

This will be needed when using the same StorageBackend instance for multiple
proxy containers.