Extract local files access check to a separate function, support .wildland-owners file
Introduce torage_backends.base.verify_local_access function that checks if
given WL user can access given local path. In addition to the old logic (user
listed in 'local-owners' setting), it also checks for .wildland-owners file in
directory to be accessed. If the file is present, it can contain a list of
allowed WL users.
This allows to give access to selected local directories to some WL users.

For now this is used only for file:// URLs.