Bear: use fixed name for the note's body
Since the note is represented as a container named by the note's title (and seen as a directory in the FS), there is no need to further repeat the note's title in the file inside the container. This approach is more compatible with possible future approach to represent each note as a TextBundle (http://textbundle.org/spec/) as hinted by Bear app itself.