Bear: Quote container title and also ensure no `/` are part of the note's title
The rationale for quoting of the note's title is that otherwise the user sees notes intermixed with tags. This makes the note's projection into the fs messy. Consider the following example:

/places/Poland/
 `- Krakow    --> a tag further refining the classification
 `- Some Note --> a leaf representing a note tagged as #places/Poland
 `- Warsaw    --> another tag refining the classification

When we prepend all the notes with a \"\ than the above situation gets a bit more tidy, as the "-prefixed dirs will typically be sorted before non-quoted dirs by most file managers.

An alternative approach would be to represent a note as a single-file, such as a TextBundle, but that would require ability to represent a Wildland container as a single file, which is currently unsupported.

The rationale for replacing `/` with '_' in the note's title is to not generate unnecessary nested dirs.