Store default user on wl start
Until wl stop, the same default user will be used that was default at the
moment wl start was executed.
Reason: avoiding confusing behavior, in which some containers are mounted with default-user that's
different from current default user. This could happen mostly if config.yaml was
changed while containers are mounted, but it was still potentially confusing.

fixes #224