Fixes problem with syncing delegated storage
DelegateProxyStorageBackend:walk() should return path relative to
subdirectory specified in delegated container's manifest, same as e.g.
DelegateProxyStorageBackend:readdir(). Instead
DelegateProxyStorageBackend:walk() was returning a path with
subdirectory path prepended. This was resulting in an exception or an
error/warning being printed, depending on the storage type pointed by a
delegated storage.

To give an insight of what was happening when user tried to sync
delegated storage pointing to a local storage, with another local
storage, the following methods were called:

  Syncer:init_state() which is calling storage.walk() to get path -->
  DelegateProxyStorageBackend:get_hash(subdir / path)             -->
  StorageBackend:get_hash(subdir / subdir / path)                 -->
  LocalStorageBackend:get_file_token(subdir / subdir / path)      -->
  os.stat(self._path(subdir / subdir / path))                     -->
  Exception thrown as there is no `subdir / subdir / path`,
  `subdir / path` should be used instead!

Closes #207