Fix `wl --help`
This commit originally was intended to address the problem with printing
backtrace when Wildland's config.yaml contains the following invalid
configuration (as described in issue #245):

  default-containers:
  # - 'this line is commented out'

This issue was addressed in commit c31cbc4 though (latest origin/master
at the time of writing). The problem is that commit c31cbc4 introduced
the bug which causes `wl --help` to print `Error: 'debug'` instead of a
help section. This commit fixes the problem in `wl --help` handling.

Fixes #245