FUSE: Don't force a directory read-only if there is just one writable storage
If there are multiple storages for a single directory, there may not be
a simple way to find what to do with new files. But if only one of those
storages is writable - that can be determined. In that case, do not
force a directory to be read-only, and also return Resolved object
pointing at that storage. This means it is no longer easy to distinguish
synthetic directories from real ones from outside of ConflictResolver
class. But if the abstraction is used properly, it should not be an
issue.

Fix #357