Implement cache storages
Cache storages are local storages used to speed up operations on slow backends.
They are explicitly defined by the user when mounting the container or by
the new container create-cache and delete-cache commands. Cache storage becomes
the primary storage for the container, its contents is synced with the old
storage when the container is mounted.