Compare just user id and pubkeys for User object equality
Do not compare infra containers, as the same container can be expressed
in different ways, leading to false duplicate results. Especially, an
inline container may have some inline-optional fields filled in or not -
making they look like a different object, while it is in fact the same.

Unfortunately, at this stage, we have bare dict with container
parameters, so we cannot make use of smart Container.__eq__.

Fix #345