search: aggressively cache search results
Each resolve step results depends only on the container where there the
resolve looks and for what path it looks for. This observation allows
caching some of the results between different search instances - for
example if multiple paths leads to the same user.
Ideally, this should make loading several WL paths, must faster, as
common parts are not re-evaluated. This is especially helpful when
mounting a forest when most containers use 'delegate' storage.