Do not re-enumerate manifests catalog when mounting the whole forest
When enumerating user's containers, manifests catalog is enumerated.
But, the same mechanism is used when looking for subcontainers. This
means, if manifests catalog is one of the containers included within it,
it is enumerated twice - once as the actual catalog, and then as a
normal container with subcontainers. This results in containers being
accessed twice - which normally is "just" a waste of time, but in some
cases may cause the mount operation to fail.

Fix this, by marking a manifests catalog, when it's loaded via this very
catalog. Then, use that flag to skip looking for its subcontainers.

Fix #335