Fix issue with Dropbox unusable as a manifests catalog
There was an issue with calling readdir() on unmounted Dropbox storage
which was resulting in an error. This issue was created as a
Dropbox-specific but probably any storage which has some
disconnect/connect logic implemented in mount/umount was affected.

To solve the problem, mounting the storage was added in search.py. This
approach didn't affect efficiency of mounting Pandora forest. At the
time of writing, Pandora mount time took 60 seconds both with this
commit and without.

Fixes #470