Transpose storage backend.
Enables the user to modify the automatically generated categories of a subcontainer by providing a set of rules for the backend to follow.
Includes 'exclude', 'include', 'replace-with' and 'match-category-regex' rules.

Added documentation for category-modification backend

added the copyright headers and return type hints

Added a helper Rule class for applying the rules to the categories.
Modified the way the rules are expected to be passed to the backend.

Added reference-container type to types.json