additional ci-test
Added a new ci-test to ensure that the backend automatically appends a trailing slash to the url entered by the user (to avoid NoSuchBucker errors when accidentally entering an invalid `s3_url`). The test also checks that the container created with the template is treated as usual and contains relevant files.