paths_only
pylint fixes
Modified the subcontainer watcher to enable for watching using paths and modification time of the files only.
modified the get_info() command to only compare paths leading to the subcontainers
added the `paths_only` argument that allows the backends to return an iterable of paths (instead of an iterable of paths and subcontainers) from the get_children method implementations
TODO: modify the handling and detecting of the SubcontainerEvents of type MODIFY so that they detect changes in the subcontainer manifests too (e.g. changes in categories)