working timeline and children paths
pylint fixes
Added simple tests for checking whether the logic behind the subcontainer remounter functions as expected
#TODO:
-modify the way the events are checked to check for both the existence and content of the files in the timeline tree
-add tests for the MODIFY event type
Modified the subcontainer watcher to enable for watching using paths and modification time of the files only.
TODO: modify the handling and detecting of the SubcontainerEvents of type MODIFY so that they detect changes in the subcontainer manifests too (e.g. changes in categories)
modified the get_info() command to only compare paths leading to the subcontainers
added the `paths_only` argument that allows the backends to return an iterable of paths (instead of an iterable of paths and subcontainers) from the get_children method implementations