docker: add HTTPS to test apache instance
This is necessary for testing WL path hints, which do not support plain
http.
For now start the server at port 443 (inside docker only) for
convenience but later it can be moved to another port. The server is
running with self-signed cert, that is added to the trusted set.

Base on work by @fepitre