s3: return the same file object if its already open for writing
POSIX requires that write() to a file is immediately visible for other
processes. Do this by returning the same S3File instance for files that
are open for writing, even if subsequent open is just for reading.

To make things easier, allow only one process to open a file for
writting at the same time. And ensure the content is uploaded only when
the writing process closes it.

Related to #801